requirements.txt
Flask
Werkzeug
Flask-Mail
python-dotenv
authlib
requests
torch
torchvision
timm
Pillow
numpy
